+++ src/api/utils/index.ts
@@ -4,3 +4,4 @@ export * from './depth';
 export * from './states';
 export * from './hash';
 export * from './uniform';
+export * from './typedarray';